\name{mded-package}
\alias{mded-package}
\docType{package}
\title{
Measuring the difference between two empirical distributions
}
\description{
The package provides a function for measuring the difference between two independent or non-independent empirical distributions and returning a significance level of the difference.
}
\section{Acknowledgments }{
I would like to thank Professor Gregory L. Poe for his kindness.
}
\author{
Hideo Aizaki
}
\references{
Poe GL, Giraud KL, Loomis JB (2005). Computational methods for measuring the difference of empirical distributions. \emph{American Journal of Agricultural Economics}, \bold{87}, 353--365.

Poe GL, Severance-Lossin EK, Welsh WP (1994). Measuring the difference (X - Y) of simulated distributions: A convolutions approach. \emph{American Journal of Agricultural Economics}, \bold{76}, 904--915.

Poe GL, Welsh MP, Champ PA (1997). Measuring the difference in mean willingness to pay when dichotomous choice contingent valuation responses are not independent. \emph{Land Economics}, \bold{73}, 255--267.
}
\note{
Recommended citations:

Aizaki H (2014). \bold{mded}: Measuring the difference between two empirical distributions, R package version 0.1-1. URL http://CRAN.R-project.org/package=mded.

Poe GL, Giraud KL, Loomis JB (2005). Computational methods for measuring the difference of empirical distributions. \emph{American Journal of Agricultural Economics}, \bold{87}, 353--365.

Poe GL, Welsh MP, Champ PA (1997). Measuring the difference in mean willingness to pay when dichotomous choice contingent valuation responses are not independent. \emph{Land Economics}, \bold{73}, 255--267.
}
\keyword{ package }

